V 0.1 Alpha Test.
Basic AI that has sanitized inputs so you don't have to worry about file problems. (do not use '  ' in the inputs as it will break)

This AI has a dedicated file that trains the AI and a file that is the UI for the chatbot.

This is CASE AND SYMBOL SENSITIVE.

Testing purposes only.