V0.2 Alpha Test
The Text and file names are now sanitized and lower case. The Alpha comes with a few tests, so your answers are no longer case sensitive. There is also a `readme.md` with important info. ANY python devs are welcome to help with the AI. The current limit for any non registry edited Windows Version has a name limit of 256 characters but in the python script, there is no hard limit (Will be in the next update).